Add brute force feature at backend with Redis to protect APIs from spamming.

+++ PetShop/server/routes/redis.js
@@ -2,7 +2,10 @@ import { Router } from "express";
 import { flushAll, flushDb } from "../controllers/redis.js";
 
 const redisRouters = Router();
-redisRouters.get("/flush-db", flushDb);
-redisRouters.get("/flush-all", flushAll);
+
+if (process.env.REACT_APP_NODE_ENV === "development") {
+  redisRouters.get("/flush-db", flushDb);
+  redisRouters.get("/flush-all", flushAll);
+}
 
 export default redisRouters;
